Font size unit. Frame filler.
Added a font size unit attribute to Wardrobe, with 'px' as the default,
because different renderers interpret the standard differently for the
case of not having a unit.

Added a simple method for filling in the space inside a frame, and made
the CardFront class apply this method, restoring opaque white bodies
for cards.